Remove duplicate log lines
This refactors the logging to introduce Formatter class that can be
used to format log lines that are written to stdout/stderr.

We introduce the following formatters:

- A basic one
- One that logs time, thread and location
- One that filters out duplicate lines.

By default we install the one that removes duplicates so the emulator
producess less spam, the previous behavior can be enabled by adding the
-log-nofilter flag.

Includes a series of tests and benchmarks to get an idea of the impact.

Bug: 214071222
Change-Id: I7142933c8c440c907177d12ba9d93837bcaea1dc