Enable RTTI on all frameworks
Some of our libraries are using RTTI, (modem, breakpad), which can
result in undefined behavior as linking RTTI against NO-RTTI is not
well defined.

Furthermore RTTI is a requirement for the TINK crypto library.

For more details and background of the issues we can encounter:
https://www.russellmcc.com/posts/2013-08-03-rtti.html

This will result in a slight size increase of the emulator (+/- 93kb)

Change-Id: I1ec1b151a4c3ca8e481f9bec204d1f5e88c482be