Add support for jwt token validation.
This adds support for JWT token validation. It adds two classes that can
be used to start validating JWT tokens:

- A JwkDirectoryObserver:

This class will observe a directory and load all the json webkeys into a
keyhandle and fire an event.

- A JwtTokenAuth class:

This class will validate that there is a valid Bearer: header with a
properly signed JWT token. The JWT token must have been signed by a Jwk
that is discovered by the directory observer.

Contains a set of unit tests to valdate the behavior.

Bug: 205713223
Change-Id: Ib53a5a0099e29ae234e5c24407707864706c9b11