shorten screenoff timeout to 30 minutes for starting with S
Based on profiling, screen off reduces idle cpu usage from
average of 16% to 12% for google apis(include playstore) images.
Setting the screen off time to 30 minutes to minimize disruption
to developer flow.

This is only for api >= S, to minimize impact. We might
revisit this and do the same to older apis later.

Bug: 216165503
Change-Id: I9b192b633cc5e7c6635a3292153595f1027440c7