Handle audio exchange case
The invariant g_actvive_sw == get_hda_voice_in does not hold. Due to
this we cannot close out g_active_sw as it might have been closed
out already.

Instead we will only close out the active voice if it is not the same
as the one we will be replacing it with.

Bug: 204380121
Change-Id: I110678e58bd7e0f1e2fe78efeb2dcfa0336f239f