[Bug Fix] Use the default DNS address in guest provided by Slirp
On Forge, the default network is IPv6 and the hard-coded 8.8.8.8 is for
IPv4. This led to internet connectivity issue when WiFi is turned off
in Google 3 tests. Instead, we can still rely on the magical DNS address
provided by Slirp and let the host decide which DNS address is best to
use.

Bug: 218162900

Signed-off-by: Weilun Du <wdu@google.com>
Change-Id: I417fbfc6aa2ccf586ed5f42dfa72eafdffc9dcff