Only process multitouch events generated from touch screen
On MacOS, we have seen un-released touch points in trackpad.
For example, Qt::TouchPointReleased is never issued for a
specific touch point, and this results in unwanted behavior
such as device unresponsiveness because Android system assumes
there is long press finger that is never moved.

With this CL, the touch point on the first press will be
interpreted as mouse press event instead by QT. In the future,
we want to properly fix the issue in multi-touch for trackpad.

Bug: 217620191
Bug: 216505499
Bug: 216383452

Signed-off-by: Weilun Du <wdu@google.com>
Change-Id: I5ebc5cb8fcd8e4ad1962056ba4b7e53849ebc88b