[gRPC] Properly handle partial closure in async
It is possible for a client to close its side when streaming. This is
only a partial closure as the server can still keep writing to the
client.

Previously we did not handle this case properly resulting in undefined
behavior (and/or crashes).

We now update the api to handle partial closures and add documentation
on how to handle partial closures properly.

Bug: 223717656
Change-Id: Ie919bf13392dcf960a8ca4b1dda1214ca19e9ae2