[gRPC] Introduce async<->sync adapter for Windows
It seems there is a bug in the gRPC engine preventing us from using the
async API for windows.

This introduces an adapter patter that adapts the sync interface to the
async interface so we can use the async interface for both.

Once the windows gRPC issues are addressed we can:

- Revert this change
- Do not set the define DISABLE_ASYNC_GRPC

Test: unit tests are now green.
Bug: 225895409
Change-Id: I23a25534cc0d086c2b8d998cdef5f8350f4f33e5