[Bluetooth] Service specification and sample
This contains a python sample demonstrating how the API is used to
define a bluetooth heart rate monitor.

It introduces a python example of how developers would define their own
emulated gatt devices, and how they would be registered with the
emulator.

Note: This sample does not yet work as the infrastructure is not
present.

Bug: 198675851
Change-Id: Ic0d626c841d5bc9afc337e7bba3529ca9630e34d