[vmnet] Configure mobile network to use vmnet if requested.
This change will enforce both the wifi and mobile network to use vmnet
if requested. This is necessary for API 30 and below because wifi and
share the same network backend.

A new cmdline option `vmnet` is added to replace `wifi-vmnet`. The old
option is kept for backward compatibility reason and it will be deleted
eventually.

Bug:  204822950
Test: run cts -m CtsWifiTestCases

Signed-off-by: Weilun Du <wdu@google.com>
Change-Id: I7b0fc7b82741f8f2e095cf7b833b089aadde7c62