Make the console agents globally available
This turns the global agents into a shared libary, making them
accessible to everyone. This will enable the refactoring of the UI
components to become independent shared libraries.

This in turn will make it possible to build two versions of the location
ui and link them dynamically.

The majority of the code is a search & replace with a clang-format
applied.

Note: This change exposed an issue with a linker update we received.
Modern linkers will setup the RUNPATH v.s. the RPATH. The RUNPATH gets
searched AFTER LD_LIBRARY_PATH, v.s. before. This causes us to load the
wrong shared library in the unit tests where we relied on the RPATH
behavior. Instead of loading libEGL.so from swiftshader we would load
the one on the main system, resulting in unexpected behavior.

This doesn't happen in the emulator since we use a launcher that
explicitly sets the LD_LIBRARY_PATH

Test: Smoke test on windows, mac, lin
Bug: 228231680
Change-Id: Iedc195ce06a1dda37344b40e220dc736898aa838