Avoid unnecessary audio resamplings
We used to advertise only one sample rate
supported (44100Hz). If Android wants to play
a 48000Hz audio, it has to resample it to
44100Hz. If the underlying hardware (on the host)
is running on a different sample rate
(e.g. 48000Hz), it has to resample 44100Hz to
48000Hz. With this change the hardware (virtio-snd)
will advertise more supported sample rates and
do resampling (if needed) on the host.

Bug: 205154312
Test: boot emulator, check if audio works
Signed-off-by: Roman Kiryanov <rkir@google.com>
Change-Id: I89c80547b8aca7481da0bef369ad888a71fc3272