Minimal android-emu library for launching emulator
We would like to create a set of shared libraries to reduce the emulator
size and decouple qemu from the UI. This means turning parts of
android-emu into a shared library.

Binplacing shared libraries on windows requires them to be in a specific
location (there's no such thing as an rpath). The emulator launcher is
used to setup all the search paths, so we can locate the shared libs
under lib64.

This means we cannot have the launcher depend on a shared library. So we
create a small library that contains all that is needed just to run the
launcher.

Bug: 228231680
Change-Id: Id8c7b7721614edfe31af65c2d7db2b779386cadb