Build protobuf as a shared library
We now build protobuf as a shared library. This will save around 100MB
in deployment and will make it possible to build the metric and
components of the UI as a shared library.

We move the tracing component out of emu-base so to make sure it does
not dependend on protobuf.

Bug: 228231680
Change-Id: I64e37aa8c12f40ee7e4612bfa7eee934f8716c2a