Move all protos to shared lib
This migrates all the protobuf libraries to a single shared library.

Protobuf libraries register themselves with the libprotobuf module
and can never be registered more than once. By moving them into
a shared module we can guarantee that the protobufs are registered
only once but can still be used from a shared library.

Bug: 230029961
Change-Id: I2dc87b84317bf67bfaac5d5311ed54283bf3a5fd