Prepare UI to become a shared module
This refactors all linker dependencies from the UI. There are now a set
of additional agent functions that further decouples the user interface
from linking against externs and other global variables.

This sets the stage for redefining the QT pieces to be standalone.

It requires a few lower layer libraries to be refactored so the modules
are not yet build as standalone modules.

Bug: 228231680
Change-Id: I47b29fe86bf10cb30448cb88019091ae5ba380fa