gl: do not override glGetTexImage on linux
The original cl aosp/2019160 was meant to only override it on Mac,
but mistakenly overrided it for all platforms.

This cl adds a compiler flag to make sure the override
only happens on Mac.

Bug: 231314183

Test:
launch S emulator image, and take a picture with camera
app. The image should be good one instead of random garbage.

Change-Id: I32466d7e1968ae716adc141fb7c0cf0cc222adc8