Fix snapshot loading for pipes that do not support loading
Not all pipes support snapshot loading, for
those we still want to save them and once one
is loaded we want to notify the kernel that
we just closed it so the kernel can notify
the userspace. This change takes care of all
other pipe types, not just the old one as in
the previous change.

Bug: 231345789
Test: load a snapshot, check if clipboard
Test: sync still works
Signed-off-by: Roman Kiryanov <rkir@google.com>
Change-Id: I61a2c27ef89806973d2690f00220776bce2c5428