[Bug Fix] emulator often doesn't respond to "call device" button on the "phone" page of "extended controls" window
This bug is caused by a flaky socket recv function that returns errno
EAGAIN or EWOULDBLOCK. Previously socketRecvAll function would return false if such error no is returned.

To fix the bug, we re-implement the ReadExact function to no longer
return false when errno EWOULDBLOCK or EAGAIN occurs. It simply
continue to read from the socket.

Bug: 232971796

Signed-off-by: Weilun Du <wdu@google.com>
Change-Id: I205a7e16cdc4205811cf7b21369fe7b3e4d806a4