[aemu] Disable version check for Fuchsia
Over the past few months, we've had multiple reports of users whose
emulators sporadically failed to start, accompanied by a message in the
log about the emulator being "out of date", and then instructing the
user to "update by launching Android Studio". This is never appropriate
for Fuchsia developers, since we don't use Android Studio, and the
emulator binary is packaged as a prebuilt inside the SDK.

This CL enables the version check only when running as Android, so
Fuchsia will always skip it.

Bug: https://bugs.fuchsia.dev/p/fuchsia/issues/detail?id=95701
Change-Id: I094a5fb0aefdd5f817bccd1dc079196035d921a0