Enable usage of various caches
This enables our cmake files to use the mozilla distributed cache.

This introduces the --ccache parameter on the build scripts that can be
used to set a compiler cache.

It has the following options:

--ccache none (or not using the flag): Use no cache
--ccache auto: Infer the best cache, sccache for windows, ccache for
  posix (if available)
--ccache MyCustomCache: Use this if you have your own cache.

For windows we introduce a workaround to handle rsp files.
sccache does not properly handle @rsp files. So instead we manually
unpack the rsp files if this will not result in a cmdline that is longer
than 32k characters.

Bug: 234476797
Change-Id: I98db66092f9e8237ed0f05e2404cea6df14e352e

tmp - will be deleted automatically

Change-Id: Ibab0ce1de20a96de709d0f384228de6a928233a8