Add a feature to send an RST frame in the vsock device on snapshot load
If a vsock device does not support loading from a snapshot,
we want to notify the guest side that the connection is
closed on the host side. Unfortunately, the clipboard
part has a problem in public S system images which causes
very bad binder issues, so we cannot send an RST frame there
until the system image side is fixed and republished.

Bug: 231345789
Test: presubmit
Signed-off-by: Roman Kiryanov <rkir@google.com>
Change-Id: I98c29d8adb27a000a14787d41a493635843d2dc0