Explicitly register rootcanal test devices
Rootcanal links using the -wholearchive flag, something that is not well
supported with cmake (it works differently on all platforms), and uses
static variables to auto register the devices (just like qemu!).

Instead of trying to support wholearchive we just explicitly register
the devices of interest. This enables us to pass the cts tests.

Test: ./objs/emulator @33 -rootcanal-default-commands-file cts_ble.cfg
Bug: 223403193
Change-Id: I3ae6c54ac0392b1e0e26fa29bea148c13313da5e