Move RingStreambuf to a separate library
This moves the Streambuffer to an indepented library that can be used by
others and adds support for closing the stream.

Closing the stream closes the stream for additional writes, and
immediately sets the timeout to 0ms unblocking all readers.

This results in readers quickly reaching eof. The addition exposed some
small performance issues around using notify_all while holding the lock
in windows.

This is needed to refactor emu-base to make betosim standalone.

Change-Id: If6f75f107570588c52e4841dcfd58ba41e88b633