Fix crash in emulator-check
If you call "emulator-check accel" it will run through a code path where
it requires FeatureImpl class. This class has the unfortunate side
effect of reading global variables.

Global variables need to be made available throught the commandline
agents.

We now inject a set of mock agents to make sure we can provide the
unused console agents, and thus preventing a crash

Includes a test that exposed the problem as part of the unit tests.

Bug: 243414921
Change-Id: I184584ae0e2e1effa4f0d70114ba1716d954a16b