Explicitly strip for linux-aarch64
We were not properly generating strip for linux-aarch64, resulting in
large unstripped binaries.

CMake does not invoke the proper objcopy command during install, so we
now explicitly strip the executables during the install step.

We now properly generate strip for linux-aarch64 and explicitly strip
the binaries during the install step.

Change-Id: I241a139890cddeb006f6048e8fcb7bf0944ee305