Add support for allow/block lists of gRPC methods
Access to the gRPC endpoint can now be protected by:

  - Enabling a gRPC token (used by studio)
  - Using a JWT token
  - Allow lists to limit scope of the tokens above.

The allow list allows the emulator owner to control which methods can be
access by whom. This will enable users to restrict access, even though
someone presents a valid token.

This will enable those who run the emulator in a shared environment
(cloud emulator) to declare which methods can be accessed and by whom.

Those running the emulator can:

- Completely open a URI, not token or auth is needed
- Only allow access when a valid token is present, the token
  does not need to present an aud filed for the URI
- Only allow access if a valid token is present and the token
  contains an aud field with the requested URI
- Deny access to a URI

See go/emu-grpc-integration for details on why we adopted this approach.
The default access list is under ../lib/emulator_access.json and will be
shipped as part of the install.

With this change we also limit the scope of the "android-studio" token.
The token will now only grant access to the methods in the "green"
section of the emulator_access.json.

Note: This does not *yet* activate this type of protection by default.
Activating the jwk authorization with also activate the studio token for
backwards compatibility with studio and gradle (icebox).

Test: A series of unit tests, a performance test to validate caching
Test: Embedded emulator works
Bug: 247612161
Bug: 217950739
Change-Id: Ia372e9b49f17423ec8033ce2a667b0276e1f56b8