Fix minimize issue on linux.
On linux, the window will not minimize if there is no minimize button on
the window. To fix this, we add back the minimize button just before
processing the minimize event. The bug comes when we setWindowFlags to
add back the minimize button, we need to manually call "show" on the
window immediately after, otherwise, Qt will show after it is minimized,
which will bring up the window again. Yuck.

Bug: 247803980
Test: Minimize with toolbar, then unminimize by clicking on window icon.
Change-Id: I8820e19beb780b99f973a2250d9a5e8dca227361