Add support for running tests
This modifies the cmake scripts so we can compile and run the rust tests
through `cargo test`.

This is done by analyzing the manifest to see if a test is declared
in the Cargo.toml. If it is we will add the test to ctest and run it
when ctest is invoked.

The test is run by creating a shell script that will invoke cargo
with the "test" instead of "build" option.

Tests will be named cargo-test_{libname}

Change-Id: If6309f22f833cd71c80fa754901801afec01e870