Enable LMP in rootcanal
This enables the lmp feature that is based on rust.
The lmp feature provides a series of improvements over the current link
layer that is used in rootcanal.

M1 is not supported due to a missing toolchain.
Windows, we need to import a crate that needs to go through security review.

This updates the corrosion cmake to add support for flags, which we use
to make sure we never download anything from the internet.

Bug: 238726856
Change-Id: I6fcc4c734156ad01bd4d35234846296d124d5fcf