Refactor and cleanup build scripts
This moves all the build related and test related scripts to python.
This makes sure we can run the same scripts on windows as posix. Adding
a series of additional tests on windows.

The build process is now divided in a series of well defined tasks,
which can be tested individually.

- Includes a set of pytests to validate behavior
- Fixes a set of small annoyances:
  - You can now create the build everywhere
  - Fix a symlink bug in tink

This re-activates the emugen tests, which appear to be broken.

You can now list the tasks with:

./android/rebuild.sh --task-list

and run a task with

./android/rebuild.sh --task IntegrationTest

you can disable a task with:

./android/rebuild.sh --task-disable clean

Bug: 258723850
Change-Id: I4d999a191f444faa09731d7ba198fc0fd287e25f