More robust building of flattenrsp.cpp on windows.
This does not work when configuring inside of VSCode
as it does not run in a visual studio command prompt,
and therefore link.exe does not exist (meaning clang_cl.exe
will not successfully create an executable).

This keeps the existing flow, but if it fails, then
it invokes vcvars64.bat and runs from there.
We already rely on the %VSINSTALLDIR% so this does not
add extra requirements there.

BUG: 259570491
TEST: Built with -DOPTION_CCACHE=.../sccache.exe in VSCode.
Change-Id: I60ae9429b9e44981a5ac5fb32a7e3060986764e3