Do not call Finish in OnCancel
The gRPC documentation is not very clear on when Finish should be
called. However it appears that calling Finish from within the OnCancel
callback results in consistent crashes.

Relevant documentation:

```
void OnCancel(): (Server only) Invoked by the library if an RPC is
canceled before it has a chance to successfully send status to the
client side. The reaction may be used for any cleanup associated with
cancellation or to guide the behavior of other parts of the system
(e.g., by setting a flag in the service logic associated with this RPC
to stop further processing since the RPC won't be able to send outbound
data anyway). Note that servers must call Finish even for RPCs that have
already been canceled as this is required to cleanup all their library
state and move them to a state that allows for calling OnDone.
```

Bug: 260968394
Change-Id: I7669c03adeca5c0ad1a61ccbde8ffbc5cc0ac2a5