Add fix task for incorrect windows checkouts
Some windows users are reporting rust compilation failures due to sha
mismatching.

This is likely due to the fact that they have configured git to do
autocrlf replacement, which results in the calculation of incorrect
shasums.

To make sure users recognize that this is happening we will color
the output when running this in an interactive terminal.

This can be fixed by running:

android\rebuild --task FixCargo
android\rebuild --task CratePrepare

This should only be needed on windows machines that run into this issue.

Change-Id: I76d1a1abd8fc817849e2bcad3e4d4dfcad3defbc