Prevent emulator from discovering itself
This improved emulator discovery by making sure the following holds:

- The pid is alive
- The pid has a gRPC endpoint
- The pid has a reachable gRPC endpoint

This makes sure we do not discover potentially dead emulator that have
a similar port configuration as ours.

This will make sure we don't accidentaly make a bluetooth mesh to
ourselves.

Adds a test to validate the improved detector and improves the logging
to make it easier to understand when a remote connection to another
emulator fails.

Bug: 262915268
Change-Id: I4b4de9bf8b231b0fad983967f90e261134265c90