Refactor whole-archive linking
We now use interfaces to enable linking of whole archive in a way that
it actually propagates to target dependencies. This will make sure this
happens on all platforms, vs. darwin only.

This makes sure that if we link a whole archive, that the dependents
will also link the whole archive.

Bug: 262886436
Change-Id: Ifa18c54d425ad8ec4cfe1fa025b4eb64e2db492b