Add support for bluetooth with packet streaming
This adds support for netsim. Netsim can be enabled by passing in the
`-packet-streamer-endpoint <uri>` flag. The emulator will revert to
using rootcanal if this flag is not prevent.

The emulator currently only supports discovery of netsim.

This deprecates the majority of the roocanal flags. Rootcanal will be
removed in one of the upcoming cls.

Note: This is not yet stable.

Test: Can discover devices from the emulator.
Bug: 255649712
Change-Id: Id568d7ed0f7f22587393aa141538aca1050e7c1f