Only set displayname if an avd has an ini file
Fuchsia is using the emulator with a different configuration. One of the
key differences is the lack of a config.ini file.

We now make sure that we only set the displayname if we actually have
a confing.ini file. This makes sure that the emulator does not crash
when gRPC is enabled and you are launching fuchsia.

Bug: 266006048
Change-Id: I3cf47451af463bbfa1666b6375c3b26df266d0ff