Enable VsockSnapshotLoadFixed_b231345789 on API33+ by default
API31 and API32 have guest side isues with clipboard after
loading from a snapshot. In addition, there were a host side
issue as well. The host side issue is fixed, but the guest
must also be fixed to enabel it. Once the guest
side issue is fixed the `VsockSnapshotLoadFixed_b231345789`
feature should be enabled on those API levels.

API33+ are fixed.

Bug: 266956109
Test: presubmit
Change-Id: I7475f23858b621185208b53acab15bc615eb40ba
Signed-off-by: Roman Kiryanov <rkir@google.com>