Update folded area right after boot complete
Folded area must be configured before the first fold command, otherwise
the guest will not update. So let's update it right after boot complete.

Note that it takes a few seconds to run through all boot complete
commands, so it is still possbile that users click the fold button
before it is updated, which will make guest unresponsive to folding.

Bug: 269183681
Change-Id: I68ef493622474ae95a84f66654c1d2db81b46a88