vulkan: Skip waiting for seqno if RenderThread receives teardown request
Recently we saw a hang on AEMU running Fuchsia where the VkDecoder
got stuck at seqno waiting. This occurs when an address space stream
is closed at the end of a process but one of its RenderThreads still
waits for seqno from another RenderThread but the other RenderThreads
have already ended running.

In such cases, it doesn't make much sense to keep the global command
sequences (because the command it waits might never arrive), but instead
the VkDecoder should just run the command to finish the cleanup.

This change introduces a new property to the VulkanProcessContext:
"skipSeqnoWait". It is set false by default, thus the VkDecoder will
wait on seqno global order by default. However if one render thread
is torn down by address space stream being destroyed, it will be set
true so that all the other RenderThreads on the same asg stream will
not do global wait since no old nor new commands are expected to arrive.

Bug: fxbug.dev/120167
Change-Id: I281d4fc23b3154991ab0b344254fe776ae313c1f