FrameBuffer: suppress createGraphcisProcessResources error message.
It seems on AEMU / FEMU rcSetPuid() may be called multiple times
and print a lot of warning messages, causing a logspam.

This just suppresses the createGraphicsProcessResources() error
logs on AEMU as a workaround. A better solution is to not call it
in rc command handlers but this requires a full cherry pick of
aosp/2258774.

Change-Id: Ic8bcf27e7635415727c546133c06275c6dc9e180