Use default python interpreter
We now use the python interpreter used to launch the build scripts on
the build bots. This makes sure we use the hermetic python interpreter
and do not have to worry about the system version getting out of sync.

If no python interpreter is provided we will fall back to the one
provided by the system.

We will no longer attempt to use python2 as this is now completely
unsupported.

Change-Id: I70e5a406c368341765be29247ba9663551717936