Skip mouse motion events when swipe gestures are synthesized.
When the guest default input device is multi-screen and the host mouse
wheel is rotated, the standalone emulator synthesizes the swipe touch
gesture.

Without the CL, the swipe geature is cancelled for the following
reasons.

1. Qt invokes EmulatorQtWindow::mouseMoveEvent after
   EmulatorQtWindow::wheelEvent.
2. In mouseMoveEvent, EmulatorQtWindow::translateMouseEventType
   translates the kEventMouseMotion event type to kEventMouseButtonUp.
3. The kEventMouseButtonUp event terminates the swipe gesture
   unintentionally.

To fix the issue, the CL skips mouseMoveEvent when the swipe gesture is
being synthesized.

Bug: 270904643
Test: Rotate the host mouse wheel
Change-Id: I68ea6c142bc02d5439940609aee839aa6ae09ede