Rename symbol artifact
The build bots will pick up dependencies with the prefix:
sdk-repo-{target}-emulator

We rename the symbol artifact to:

`sdk-repo-{target}-emulator-symbols-xxxx.zip

so it will be picked up by the build bots and made available to the test
targets.

This should make the symbols available to the e2e tests.

Bug: 271168786
Change-Id: Ie79164f738d14af71d458a695f4066dd9890d380