Re-enable netsim in cross compile
The mac build was broken due to undefined behavior in the
android_link_libraries function.

This function does not always behave as expected and in certain
conditions it could provide a set of arguments that did not exist,
resulting in additional libraries being linked that were unused.

In the case of the cross-compile this could result in linking against an
API key for example (for example -l39432). This in turn would break the
build.

Since cross compilation does not produce any output you would not have
any idea what is going wrong.

With this change we:

- Add android_target_link_libraries2 that does not have this issue
- Clean up the breakpad cmakelist to only use what we need
- Move the symbols function to its own cmake file.

Bug: 271329265
Change-Id: I192b5aae3b8957be5801890da02b0e012c3902e0