Migrate to libyuv from webrtc
We compile the libyuv library twice on some platforms. We now migrate to
use the one that ships with the WebRTC branch. This bumps the version
from 1652 -> 1856

This makes sure that we only link against one libyuv version, that is
made available through webrtc.

This upgrade introduces +/- 3k of unittests and contains compiler
specific optimizations for our supported platforms.

Change-Id: I439fa8d260b26fcc5c68bff78f5079eeb085c1de