Reduce logging spam from test failures
This binplaces all the logging xml files into the testlog directory, so
it can be picked up by our test infrastructure if we enable that.

We also reduce the logging spam from failures in general. We now only
output tests from failed runs, versus outputting the results of all the
tests.

This should reduce the amount of logging spam we see in the build.log
file.

Change-Id: I6fe7ba106d18658a5c3dae31c0d72f873adf9b44