Fix build break due to boringssl api change
This commit fixes a build error caused by a change in the BoringSSL
library's public API. Specifically, the commit with ID
890c201d4ac9c345c304d646365fe077cf2b60c1 made the EVP_PKEY opaque, which
means that code that accessed the EVP_PKEY struct directly will no
longer compile.

To address this issue, we followed the update note provided with the
change, which recommends using accessors instead of directly accessing
the EVP_PKEY struct. The most common fix is to use EVP_PKEY_id(pkey) for
pkey->type.

Bug: 273801294
Change-Id: I2f5516f14a2ac377544e216cf394ed62252996c1