Add JwkKeyLoader class and unit tests
This commit adds the JwkKeyLoader class, which can load JSON Web Key
(JWK) sets from files or strings and manage the active set of keys. The
class provides methods to add, remove, and retrieve JWK sets, as well as
to clear and check the size of the collection.

Unit tests are also included to ensure the functionality of the class.

This is now used by the directory observer to manage they keys found in
a directory.

Bug: 273331311
Change-Id: Icaf492aab3110eb2aa2d949388e1c7203bf0a4e8