Improve Windows change detection
This commit improves the change detection iterator on Windows by fixing
a bug that caused the pointer to not move forward correctly. The bug
resulted in duplicate events being fired, which prevented the program
from detecting multiple events. Unfortunately, the existing unit tests
only validated the arrival of a single event, which did not catch this
issue. To address this, a new test that exposes the issue was added.

Bug: 273331311
Change-Id: Ia8aa3731d753efcc9e4a3fb3ad0e96ad592a714c