Add retries for crashreporter
The crahsreporter tool needs to obtain the crash database. However it is
possible that the crash database is still locked due to another process
that is still in the process of crashing.

We will now attempt a retry if we cannot obtain the crash database due
to a locking issues

We also add a secret developer option that allows you to dump crashes
if you are below the objs directory.

Bug: 273992302
Change-Id: I2f1295c21a11e16a815e8add26ced449d4f38034