Enable Nimble to connect to remote netsim
Add support for connecting to a remote netsim address set in the
environment variable NETSIM_GRPC_ADDRESS. If the environment variable is
not set, the code will continue to use the default address.

This change improves the flexibility of the code by allowing it to
connect to a remote netsim endpoint if needed.

Test: Able to connect to remote netsim server
Change-Id: If70a39c2747be572bb628f40c53218d6df8d3b1f