[gRPC] Woraround for write delays on gWindows
We are seeing write delays on gWindows where files are not immediately
available. To work around this issue we now will attempt to retry
loading the JWK if the file size is reported to be empty.

We will attempt a few times, waiting at most 1 second.

Includes a unit test to validate the behavior.

Bug: 274984168
Change-Id: I7715fbc80a043bff11a6903fe886c577ca928308