Add sensor service
This separates the sensor services and makes it compliant with protobuf
language requirements. And includes support for event streaming.

This can be used to migrate the the virtual-sensors-page.cpp to use
gRPC.

Fixes a bug with the generalized stream support where writers were not
closed in case of stream cancellation.

Enables the incubating services in the python module for usage in e2e tests.

Bug: 268207875
Change-Id: Ife4a97c9bfa42793495b162b3f3735a94d73fda5