hw/i3c: Split DesignWare I3C out of Aspeed I3C
The Aspeed I3C IP block is technically an Aspeed IP block that managed
6 DW I3C controllers.

To help reflect this better and to make it easier for other SoCs to use
the DW I3C model, we'll split out the DW portion from the Aspeed
portion.

As part of this we also do some minor changes to the DesignWare
initialization.
We make the vmstate description extern so other models (like Aspeed) can
use it.
We add sysbus_mmio_init, since it's now standalone.

Google-Bug-Id: 245567527
Signed-off-by: Joe Komlodi <komlodi@google.com>
Change-Id: I2ec2b7f42e2f3ad768fdbd01f671bf7e7ee40936

qtest/remote-i3c: Revive the test

This was either dropped on accident, or I forgot to add it in the first
place.
Either way, this adds the Remote I3C qtest, and removes some bitrot,
since it needed to be updated for the Aspeed+DesignWare I3C model split.

Signed-off-by: Joe Komlodi <komlodi@google.com>
Change-Id: Ib5fe56fc14f65801bf481ee1fb33f1a503ea0179