hw/i3c/dw-i3c: Add object props for config parameters
Since we're factoring the Aspeed device controllers into the DesignWare
controllers, we should also abstract out some of the configuration
parameters.

This allows people to instantiate their DW I3C model with their
appropriate configuration parameters instead of having to use the
configuration that ships with AST26xx BMCs.

As part of this we also move the device's object properties into a
config struct in order to help organize them.

Google-Rebase-Count: 2
Google-Bug-Id: 245567527
Signed-off-by: Joe Komlodi <komlodi@google.com>
Change-Id: Idc0e499926a84cacb60e13665fe19ee807247a56