[android-wifi-service] Switch to libslirp
libslirp is used by the latest version of QEMU and currently our QEMU
codebase won't be able to interoperate with libslirp wihtout major
rewrite and we have plans to migrate to newer version of QEMU.
The stop-gap solution is to have libslirp in android-wifi-service
and only netsim target includes android-wifi-service. Therefore,
emulator targets does not depend on libslirp.

BUG: 255961329

Change-Id: I3c0d18f902fea3da915b2bf0013811960a18a380
Signed-off-by: Weilun Du <wdu@google.com>